feat: Phase5 — フロント切り替え・Next.js API Route削除
- 全fetchコールを NEXT_PUBLIC_API_URL ベースに変更（API_BASE定数）
- src/app/api/ を完全削除（Go API サーバーに移行済み）
- サーバーサイド専用 lib を削除（db, ollama, embeddings, rss-parser 等）
- SnapshotMeta/FeedGroupWithItems/GroupInspectDetail を src/types へ移動
- 削除ライブラリのテスト3本を削除
- .env.local.sample に NEXT_PUBLIC_API_URL を追加
