+++ .env.local.sample
@@ -1,3 +1,7 @@
+# ── Go API サーバー ───────────────────────────────────────
+# Go API サーバーのURL（空の場合は Next.js の相対パスを使用）
+NEXT_PUBLIC_API_URL=http://localhost:8091
+
 # ── Ollama ────────────────────────────────────────────────
 # OllamaサーバーのベースURL
 OLLAMA_BASE_URL=http://localhost:11434
