+++ src/lib/api-url.ts
@@ -1,0 +1,1 @@
+export const API_BASE = process.env.NEXT_PUBLIC_API_URL ?? "";
